backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
filetype==1.2.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-core
langchain-openai # Changed from google-genai
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
motor==3.7.1
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
pyparsing==3.2.5
pypdf==6.4.1
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
openai==1.40.0 # Added OpenAI
